+++ main_project/component/voltageControlledSource.hpp
@@ -11,16 +11,15 @@
 */
 class VoltageControlledSource: public Component{
 protected:
-    double gain;
+    float gain;
 
 public:
-    VoltageControlledSource(string _name, vector<string> args, vector<double> extraInfo);
-    // VoltageControlledSource(string _name, double _voltage, int n1, int n2, int nc1, int nc2);
+    VoltageControlledSource(string _name, vector<string> args, vector<float> extraInfo);
     virtual ~VoltageControlledSource() = 0;
 
     void setupBasic(int n1, int n2, int nc1, int nc2);
 
-    double getGain() const override;
+    float getGain() const override;
 
     //returns all 4 nodes (N+, N-, NC+, NC-)
     vector<int> getNodes() const override;
